Many major changes. Added database, worked on chord editing, added pausing/resuming
Added MySQL support. Still need to add cloud support.
Moving toward a database system rather than a file system.
When a user clicks a chord and is presented with the editing interface, clicking the new button "confirm" sends the updated notes (as well as the location of the chord in the song/composition) to a post request.
Added composition/measure/chord tables with test entries to the database, which are then correctly parsed and outputted to the console.
Implemented pausing, resuming, and restarting mechanics, but still need to implement skipping.
Currently working on a beam that moves across the composition as the song is played.