Chords can now be edited to an extent. Database is updated to reflect edits
New measures (pre-populated with rests) can be added to the composition with no errors.
Chords can be edited by the user, and their notes can be changed. The database is updated accordingly to reflect this and the site automatically updates. (I still need to implement duration editing and the functionality to create a chord in the measure itself.)
A measure object (which points to all measures in the song/composition) can now be created from the database via the measure and chord tables. It is sent as an attribute to display the composition on the site.